do real remove expire repo image

+++ gitrepo.go
@@ -49,11 +49,10 @@ func fetchMirrorFromRemoteUnshallow(repository string) {
 	//remove expire repo  only make a log , do real remove at next commit!!!!!
 	if CacheIsExpire(remote) {
 		log.Printf("!!!!!!!!remove expire repo: %s\n", remote)
-		/*if strings.Contains(local, g_Basedir) && !(local == g_Basedir) {
+		if strings.Contains(local, g_Basedir) && !(local == g_Basedir) {
 			os.RemoveAll(local)
 		}
 		return
-		*/
 	}
 	log.Printf("git remote update: %s begin\n", local)
 	err := fetchMirrorFromRemote(remote, local, "update")
